fix(ssg): avoid circular dependency when using localstorage
